4-digit number guessing
This program is designed to guess a FOUR digit number with no repeated digits.
So, choose a “secret number” with no repeated digits and have it written down to let me guess.

Also, you need to give me two pieces of information about my guess number compared with your secret number after each guess.
One is how many digits in my guess numbers are in your secret number, and we call them to be 'Correct'.
The other is how many digits in my guess numbers are exactly in the same place as your secret number, and we call them to be 'Exact'